fixed functions in admins.js and added category

+++ graphql/resolvers/admins.js
@@ -12,6 +12,7 @@ const Question = require("../../models/Question");
 const QuestionTemplate = require("../../models/QuestionTemplate");
 const Module = require("../../models/Module");
 const Challenge = require("../../models/Challenge");
+const Category = require("../../models/Category");
 
 const checkAdminAuth = require("../../util/checkAdminAuth");
 function generateToken(admin) {
@@ -133,7 +134,7 @@ module.exports = {
         throw AuthenticationError;
       }
 
-      targetQuestion = Question.findOne({
+      const targetQuestion = Question.findOne({
         questionDescription: questionDescription,
       });
 
@@ -170,8 +171,8 @@ module.exports = {
         throw AuthenticationError;
       }
 
-      targetQuestionTemplate = QuestionTemplate.findOne({
-        inputFields: inputFields,
+      const targetQuestionTemplate = QuestionTemplate.findOne({
+        inputFields,
       });
 
       if (targetQuestionTemplate === null) {
@@ -200,11 +201,11 @@ module.exports = {
         throw AuthenticationError;
       }
 
-      targetModule = Module.findOne({
-        name: name,
+      const targetModule = Module.findOne({
+        name,
       });
 
-      if (targetModule !== null) {
+      if (targetModule === null) {
         const newModule = new Module({
           name,
           type,
@@ -213,7 +214,7 @@ module.exports = {
           createdAt: new Date(),
         });
 
-        targetAdmin.modules.push(newModule);
+        await targetAdmin.modules.push(newModule.id);
         await targetAdmin.modules.save();
         const updatedModules = await targetAdmin.modules;
         return updatedModules;
@@ -234,19 +235,50 @@ module.exports = {
       } catch (error) {
         throw AuthenticationError;
       }
-      const newChallenge = new Challenge({
-        questionDescription,
-        categoryId,
-        image,
-        createdAt: new Date(),
-      });
+      const targetChallenge = Challenge.findOne({ questionDescription });
+      if (targetChallenge === null) {
+        const newChallenge = new Challenge({
+          questionDescription,
+          categoryId,
+          image,
+          createdAt: new Date(),
+        });
 
-      targetAdmin.challenges.push(newChallenge);
-      await targetAdmin.challenges.save();
-      const updatedChallenges = await targetAdmin.challenges;
-      return updatedChallenges;
+        await targetAdmin.challenges.push(newChallenge.id);
+        await targetAdmin.challenges.save();
+        const updatedChallenges = await targetAdmin.challenges;
+        return updatedChallenges;
+      } else {
+        const unchangedChallenges = await targetAdmin.challenges;
+        return unchangedChallenges;
+      }
     },
 
+    async createNewCategory(_, { name, modules, challenges }, context) {
+      try {
+        const admin = checkAdminAuth(context);
+        var targetAdmin = await Admin.findById(admin.id);
+      } catch (error) {
+        throw AuthenticationError;
+      }
+      const targetCategory = Category.findOne({ name });
+      if (targetCategory !== null) {
+        const unchangedCategories = await targetAdmin.categories;
+        return unchangedCategories;
+      } else {
+        const newCategory = new Category({
+          name,
+          modules,
+          challenges,
+          createdAt: new Date(),
+        });
+
+        await targetAdmin.categories.push(newCategory.id);
+        await targetAdmin.categories.save();
+        const updatedCategories = await targetAdmin.categories;
+        return updatedCategories;
+      }
+    },
     async editModule(
       _,
       { moduleId, newName, newType, newCategoryId, newFormat },
@@ -259,14 +291,17 @@ module.exports = {
         throw AuthenticationError;
       }
       var targetModule = await targetAdmin.modules.findById(moduleId);
+      if (targetModule === null) {
+        throw UserInputError("Invalid input");
+      } else {
+        targetModule.name = newName;
+        targetModule.categoryId = newCategoryId;
+        targetModule.format = newFormat;
+        targetModule.type = newType;
 
-      targetModule.name = newName;
-      targetModule.categoryId = newCategoryId;
-      targetModule.format = newFormat;
-      targetModule.type = newType;
-
-      await targetModule.save();
-      return targetModule;
+        await targetModule.save();
+        return targetModule;
+      }
     },
 
     async editQuestion(
@@ -314,13 +349,15 @@ module.exports = {
       var targetQuestionTemplate = await targetAdmin.modules.questions.findById(
         questionTemplateId
       );
-
-      targetQuestionTemplate.categoryId = newCategory;
-      targetQuestionTemplate.inputFields = newInputFields;
-      targetQuestionTemplate.type = newType;
-      await targetQuestionTemplate.save();
-
-      return targetQuestionTemplate;
+      if (targetQuestionTemplate === null) {
+        throw UserInputError("Invalid input");
+      } else {
+        targetQuestionTemplate.categoryId = newCategory;
+        targetQuestionTemplate.inputFields = newInputFields;
+        targetQuestionTemplate.type = newType;
+        await targetQuestionTemplate.save();
+        return targetQuestionTemplate;
+      }
     },
 
     async editChallenge(
@@ -335,15 +372,38 @@ module.exports = {
         throw AuthenticationError;
       }
       var targetChallenge = await Challenge.findById(challengeId);
-
-      targetChallenge.categoryId = newCategoryId;
-      targetChallenge.questionDescription = newquestionDescription;
-      targetChallenge.image = newImage;
-
-      await targetChallenge.save();
-      return targetChallenge;
+      if (targetChallenge === null) {
+        throw UserInputError("Invalid input");
+      } else {
+        targetChallenge.categoryId = newCategoryId;
+        targetChallenge.questionDescription = newquestionDescription;
+        targetChallenge.image = newImage;
+        await targetChallenge.save();
+        return targetChallenge;
+      }
     },
-
+    async editCategory(
+      _,
+      { categoryId, newName, newModules, newChallenges },
+      context
+    ) {
+      try {
+        const admin = checkAdminAuth(context);
+        var targetAdmin = await Admin.findById(admin.id);
+      } catch (error) {
+        throw AuthenticationError;
+      }
+      var targetCategory = await Category.findById(categoryId);
+      if (targetCategory === null) {
+        throw UserInputError("Invalid input");
+      } else {
+        targetCategory.name = newName;
+        targetCategory.modules = newModules;
+        targetCategory.challenges = newChallenges;
+        await targetCategory.save();
+        return targetCategory;
+      }
+    },
     async deleteModule(_, { moduleId }, context) {
       try {
         const admin = checkAdminAuth(context);
@@ -352,10 +412,14 @@ module.exports = {
         throw AuthenticationError;
       }
       const targetModule = await targetAdmin.modules.findById(moduleId);
-      await targetModule.delete();
+      if (targetModule === null) {
+        throw UserInputError("Invalid input");
+      } else {
+        await targetModule.delete();
 
-      const updatedModules = await targetAdmin.modules;
-      return updatedModules;
+        const updatedModules = await targetAdmin.modules;
+        return updatedModules;
+      }
     },
 
     async deleteQuestion(_, { questionId }, context) {
@@ -368,10 +432,14 @@ module.exports = {
       const targetQuestion = await targetAdmin.modules.questions.findById(
         questionId
       );
-      await targetQuestion.delete();
-
-      const updatedQuestions = await targetAdmin.modules.questions;
-      return updatedQuestions;
+      if (targetQuestion === null) {
+        throw UserInputError("Invalid input");
+      } else {
+        await targetQuestion.delete();
+        await targetAdmin.modules.questions.save();
+        const updatedQuestions = await targetAdmin.modules.questions;
+        return updatedQuestions;
+      }
     },
 
     async deleteQuestionTemplate(_, { questionTemplateId }, context) {
@@ -384,10 +452,14 @@ module.exports = {
       const targetQuestionTemplate = await targetAdmin.questionTemplates.findById(
         questionTemplateId
       );
-      await targetQuestionTemplate.delete();
+      if (targetQuestionTemplate === null) {
+        throw UserInputError("Invalid input");
+      } else {
+        await targetQuestionTemplate.delete();
 
-      const updatedQuestionTemplates = await targetAdmin.questionTemplates;
-      return updatedQuestionTemplates;
+        const updatedQuestionTemplates = await targetAdmin.questionTemplates;
+        return updatedQuestionTemplates;
+      }
     },
     async deleteChallenge(_, { challengeId }, context) {
       try {
@@ -396,13 +468,34 @@ module.exports = {
       } catch (error) {
         throw AuthenticationError;
       }
-      const targetChallenge = await targetAdmin.Challenges.findById(
+      const targetChallenge = await targetAdmin.challenges.findById(
         challengeId
       );
-      await targetChallenge.delete();
-
-      const updatedChallenges = await targetAdmin.challenges;
-      return updatedChallenges;
+      if (targetChallenge === null) {
+        throw UserInputError("Invalid input");
+      } else {
+        await targetChallenge.delete();
+        const updatedChallenges = await targetAdmin.challenges;
+        return updatedChallenges;
+      }
+    },
+    async deleteCategory(_, { categoryId }, context) {
+      try {
+        const admin = checkAdminAuth(context);
+        var targetAdmin = await Admin.findById(admin.id);
+      } catch (error) {
+        console.log(error);
+        throw AuthenticationError;
+      }
+      const targetCategory = await targetAdmin.categories.findById(categoryId);
+      if (targetCategory === null) {
+        throw UserInputError("Invalid input");
+      } else {
+        await targetCategory.delete();
+        await targetAdmin.categories.save();
+        const updatedCategories = await targetAdmin.categories;
+        return updatedCategories;
+      }
     },
   },
 };
